requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.6.15
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
distro==1.9.0
Django==5.2.3
et_xmlfile==2.0.0
fonttools==4.42.0
ftfy==6.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
kiwisolver==1.4.4
mysqlclient==2.2.7
numpy==1.25.2
openai==1.91.0
openpyxl==3.1.5
packaging==23.1
pandas==2.3.0
Pillow==10.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.0.9
pypdf==5.6.1
python-dateutil==2.8.2
python-dotenv==1.1.1
pytz==2025.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typhoon-ocr==0.3.8
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
wcwidth==0.2.13
